+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This file is used to list changes made in each version of cookbook-openstack-ops
 * Configure rabbitmq port for both ssl and non-ssl cases
 * Fix metadata version constraint for common
 * Bump Chef gem to 11.16
+* No need to push our rabbit user/password to rabbit cookbook
 
 ## 10.0.0
 * Upgrading to Juno